refactor: Global 패키지 하위 디렉토리 구조 및 명명 변경
