+++ src/main/java/com/devpass/domain/devexperience/service/DevExperienceService.java
@@ -5,8 +5,8 @@ import com.devpass.domain.devexperience.dto.request.DevExperienceAddRequest;
 import com.devpass.domain.devexperience.dto.response.DevExperienceResponseDTO;
 import com.devpass.domain.devexperience.entity.DevExperience;
 import com.devpass.domain.devexperience.repository.DevExperienceRepository;
-import com.devpass.global.error.ErrorCode;
-import com.devpass.global.error.exception.BusinessException;
+import com.devpass.global.payload.apicode.ErrorCode;
+import com.devpass.global.payload.error.exception.BusinessException;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
