+++ src/main/java/com/devpass/domain/favorite/entity/Favorite.java
@@ -2,7 +2,7 @@ package com.devpass.domain.favorite.entity;
 
 import com.devpass.domain.recruitment.entity.Recruitment;
 import com.devpass.domain.user.entity.User;
-import com.devpass.global.common.BaseEntity;
+import com.devpass.global.common.entity.BaseEntity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
