+++ src/main/java/com/devpass/domain/internship/controller/InternshipController.java
@@ -3,8 +3,8 @@ package com.devpass.domain.internship.controller;
 import com.devpass.domain.internship.dto.request.InternshipAddRequest;
 import com.devpass.domain.internship.dto.response.InternshipResponseDTO;
 import com.devpass.domain.internship.service.InternshipService;
-import com.devpass.global.common.response.CustomResponse;
-import com.devpass.global.result.ResultCode;
+import com.devpass.global.payload.wrapper.ApiResponse;
+import com.devpass.global.payload.apicode.SuccessCode;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -22,10 +22,10 @@ public class InternshipController {
             description = "개발 경험 등록 페이지 - 인턴십 경험 등록(개발 경험 id를 기준으로 등록)"
     )
     @PostMapping("/{devExperience_id}")
-    public CustomResponse<InternshipResponseDTO> addInternship(
+    public ApiResponse<InternshipResponseDTO> addInternship(
             @PathVariable("devExperience_id") Long devExperienceId,
             @RequestBody InternshipAddRequest request) {
         internshipService.addInternship(devExperienceId, request);
-        return CustomResponse.of(ResultCode.CREATED);
+        return ApiResponse.of(SuccessCode.CREATED);
     }
 }
