+++ src/main/java/com/devpass/domain/internship/service/InternshipService.java
@@ -7,8 +7,8 @@ import com.devpass.domain.internship.dto.request.InternshipAddRequest;
 import com.devpass.domain.internship.dto.response.InternshipResponseDTO;
 import com.devpass.domain.internship.entity.Internship;
 import com.devpass.domain.internship.repository.InternshipRepository;
-import com.devpass.global.error.ErrorCode;
-import com.devpass.global.error.exception.BusinessException;
+import com.devpass.global.payload.apicode.ErrorCode;
+import com.devpass.global.payload.error.exception.BusinessException;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
