+++ src/main/java/com/devpass/domain/project/controller/ProjectController.java
@@ -3,8 +3,8 @@ package com.devpass.domain.project.controller;
 import com.devpass.domain.project.dto.request.ProjectAddRequest;
 import com.devpass.domain.project.dto.response.ProjectResponseDTO;
 import com.devpass.domain.project.service.ProjectService;
-import com.devpass.global.common.response.CustomResponse;
-import com.devpass.global.result.ResultCode;
+import com.devpass.global.payload.wrapper.ApiResponse;
+import com.devpass.global.payload.apicode.SuccessCode;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -24,17 +24,17 @@ public class ProjectController {
     )
     // 프로젝트 등록 API
     @PostMapping("/{devExperience_id}")
-    public CustomResponse<ProjectResponseDTO> addProject(
+    public ApiResponse<ProjectResponseDTO> addProject(
             @PathVariable("devExperience_id") Long devExperienceId,
             @RequestBody ProjectAddRequest request) {
         ProjectResponseDTO projectResponseDTO = projectService.addProject(devExperienceId, request);
-        return CustomResponse.of(ResultCode.CREATED, projectResponseDTO);
+        return ApiResponse.of(SuccessCode.CREATED, projectResponseDTO);
     }
 
     // 프로젝트 조회 API
     @GetMapping("/{project_id}")
-    public CustomResponse<ProjectResponseDTO> getProject(@PathVariable("project_id") Long id) {
+    public ApiResponse<ProjectResponseDTO> getProject(@PathVariable("project_id") Long id) {
         ProjectResponseDTO projectResponseDTO = projectService.getProjectById(id);
-        return CustomResponse.of(ResultCode.OK, projectResponseDTO);
+        return ApiResponse.of(SuccessCode.OK, projectResponseDTO);
     }
 }
