+++ src/main/java/com/devpass/domain/project/service/ProjectService.java
@@ -7,8 +7,8 @@ import com.devpass.domain.project.dto.request.ProjectAddRequest;
 import com.devpass.domain.project.dto.response.ProjectResponseDTO;
 import com.devpass.domain.project.entity.Project;
 import com.devpass.domain.project.repository.ProjectRepository;
-import com.devpass.global.error.ErrorCode;
-import com.devpass.global.error.exception.BusinessException;
+import com.devpass.global.payload.apicode.ErrorCode;
+import com.devpass.global.payload.error.exception.BusinessException;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
