+++ src/main/java/com/devpass/domain/resume/controller/ResumeController.java
@@ -2,8 +2,8 @@ package com.devpass.domain.resume.controller;
 
 import com.devpass.domain.resume.document.ResumeDocument;
 import com.devpass.domain.resume.service.ResumeService;
-import com.devpass.global.common.response.CustomResponse;
-import com.devpass.global.result.ResultCode;
+import com.devpass.global.payload.wrapper.ApiResponse;
+import com.devpass.global.payload.apicode.SuccessCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,16 +18,16 @@ public class ResumeController {
     private final ResumeService resumeService;
 
     @GetMapping("/devExprience/{devExperience_id}")
-    public CustomResponse<ResumeDocument> generateResume(@PathVariable("devExperience_id") Long devExperienceId) {
+    public ApiResponse<ResumeDocument> generateResume(@PathVariable("devExperience_id") Long devExperienceId) {
         ResumeDocument resume = resumeService.generateAndSaveResume(devExperienceId);
-        return CustomResponse.of(ResultCode.OK, resume);
+        return ApiResponse.of(SuccessCode.OK, resume);
     }
 
 
     @GetMapping("/{resume_id}")
-    public CustomResponse<ResumeDocument> getResume(@PathVariable("resume_id") String resumeId) {
+    public ApiResponse<ResumeDocument> getResume(@PathVariable("resume_id") String resumeId) {
         ResumeDocument resume = resumeService.getResumeById(resumeId);
-        return CustomResponse.of(ResultCode.OK, resume);
+        return ApiResponse.of(SuccessCode.OK, resume);
     }
 
 }
