+++ src/main/java/com/devpass/domain/resume/entity/Resume.java
@@ -1,7 +1,7 @@
 package com.devpass.domain.resume.entity;
 
 import com.devpass.domain.user.entity.User;
-import com.devpass.global.common.BaseEntity;
+import com.devpass.global.common.entity.BaseEntity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
