--- src/main/java/com/devpass/global/common/response/ReasonDto.java
@@ -1,13 +1,0 @@
-package com.devpass.global.common.response;
-
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@Builder
-public class ReasonDto {
-    private HttpStatus status;
-    private String code;
-    private String message;
-}
