--- src/main/java/com/devpass/global/common/status/BaseStatus.java
@@ -1,7 +1,0 @@
-package com.devpass.global.common.status;
-
-import com.devpass.global.common.response.ReasonDto;
-
-public interface BaseStatus {
-    public ReasonDto getReason();
-}
