--- src/main/java/com/devpass/global/error/exception/BusinessException.java
@@ -1,32 +1,0 @@
-package com.devpass.global.error.exception;
-
-import com.devpass.global.error.ErrorCode;
-import com.devpass.global.error.ErrorResponse;
-import com.devpass.global.error.ErrorResponse.FieldError;
-import java.util.ArrayList;
-import java.util.List;
-import lombok.Getter;
-
-
-@Getter
-public class BusinessException extends RuntimeException{
-
-    private ErrorCode errorCode;
-    private List<FieldError> errors = new ArrayList<>();
-
-    public BusinessException(String message, ErrorCode errorCode){
-        super(message);
-        this.errorCode = errorCode;
-    }
-
-    public BusinessException(ErrorCode errorCode){
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public BusinessException(ErrorCode errorCode, List<ErrorResponse.FieldError> errors){
-        super(errorCode.getMessage());
-        this.errors = errors;
-        this.errorCode = errorCode;
-    }
-}
