--- src/main/java/com/devpass/global/error/exception/EntityNotFoundException.java
@@ -1,10 +1,0 @@
-package com.devpass.global.error.exception;
-
-import com.devpass.global.error.ErrorCode;
-
-public class EntityNotFoundException extends BusinessException {
-
-    public EntityNotFoundException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
