+++ src/main/java/com/devpass/global/payload/advice/ApiResponseStatusSetter.java
@@ -1,0 +1,32 @@
+package com.devpass.global.payload.advice;
+
+import org.springframework.core.MethodParameter;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.server.ServerHttpRequest;
+import org.springframework.http.server.ServerHttpResponse;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice;
+
+import com.devpass.global.payload.wrapper.ApiResponse;
+
+@RestControllerAdvice
+public class ApiResponseStatusSetter implements ResponseBodyAdvice<Object> {
+
+	@Override
+	public boolean supports(MethodParameter returnType, Class converterType) {
+		return true;
+	}
+
+	@Override
+	public Object beforeBodyWrite(Object body, MethodParameter returnType, MediaType selectedContentType,
+		Class selectedConverterType, ServerHttpRequest request, ServerHttpResponse response) {
+		if (body instanceof ApiResponse) {
+			ApiResponse<?> customResponse = (ApiResponse<?>)body;
+			HttpStatus status = customResponse.getHttpStatus();
+			response.setStatusCode(status);
+		}
+
+		return body;
+	}
+}
