+++ src/main/java/com/devpass/global/payload/apicode/ErrorCode.java
@@ -1,0 +1,45 @@
+package com.devpass.global.payload.apicode;
+
+import org.springframework.http.HttpStatus;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum ErrorCode implements ResponseCode {
+
+	// Global
+	INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "COMMON_500", "서버 에러입니다. 관리자에게 문의하세요."),
+	BAD_REQUEST(HttpStatus.BAD_REQUEST, "COMMON_400", "잘못된 요청입니다."),
+	UNAUTHORIZED(HttpStatus.UNAUTHORIZED, "COMMON_401", "인증이 필요합니다."),
+	FORBIDDEN(HttpStatus.FORBIDDEN, "COMMON_403", "금지된 요청입니다."),
+	NOT_FOUND(HttpStatus.NOT_FOUND, "COMMON_404", "찾을 수 없는 요청입니다."),
+	METHOD_NOT_ALLOWED(HttpStatus.METHOD_NOT_ALLOWED, "COMMON_405", "허용되지 않은 메소드입니다."),
+	INVALID_REQUEST(HttpStatus.BAD_REQUEST, "COMMON_400", "잘못된 요청입니다."),
+	INPUT_VALUE_INVALID(HttpStatus.BAD_REQUEST, "REQUEST_400", "요청사항에 필수 인자가 누락되었습니다"),
+	HTTP_MESSAGE_NOT_READABLE(HttpStatus.BAD_REQUEST, "G005", "request message body가 없거나, 값 타입이 올바르지 않습니다."),
+
+	// Recruitment
+	RECRUITMENT_NOT_FOUND(HttpStatus.NOT_FOUND, "RECRUITMENT_404", "해당 채용 정보를 찾을 수 없습니다."),
+
+	GPT_RESPONSE_PARSE_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "GPT_PARSE_500", "GPT 응답 파싱에 실패했습니다."),
+
+	RESUME_NOT_FOUND(HttpStatus.NOT_FOUND, "RESUME_404", "이력서를 찾을 수 없습니다."),
+	// Company
+	COMPANY_NOT_FOUND(HttpStatus.NOT_FOUND, "COMPANY_404", "해당 회사 정보를 찾을 수 없습니다."),
+	;
+
+	private final HttpStatus httpStatus;
+	private final String code;
+	private final String message;
+
+	@Override
+	public Reason getReason() {
+		return Reason.builder()
+			.status(httpStatus)
+			.code(this.code)
+			.message(this.message)
+			.build();
+	}
+}
