+++ src/main/java/com/devpass/global/payload/apicode/ResponseCode.java
@@ -1,0 +1,5 @@
+package com.devpass.global.payload.apicode;
+
+public interface ResponseCode {
+	Reason getReason();
+}
