+++ src/main/java/com/devpass/global/payload/error/exception/EntityNotFoundException.java
@@ -1,0 +1,9 @@
+package com.devpass.global.payload.error.exception;
+
+import com.devpass.global.payload.apicode.ErrorCode;
+
+public class EntityNotFoundException extends BusinessException {
+	public EntityNotFoundException(ErrorCode errorCode) {
+		super(errorCode);
+	}
+}
