--- src/main/java/com/devpass/global/result/ResponseInterceptor.java
@@ -1,31 +1,0 @@
-package com.devpass.global.result;
-
-import com.devpass.global.common.response.CustomResponse;
-import org.springframework.core.MethodParameter;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.server.ServerHttpRequest;
-import org.springframework.http.server.ServerHttpResponse;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice;
-
-@RestControllerAdvice
-public class ResponseInterceptor implements ResponseBodyAdvice<Object> {
-
-    @Override
-    public boolean supports(MethodParameter returnType, Class converterType) {
-        return true;
-    }
-
-    @Override
-    public Object beforeBodyWrite(Object body, MethodParameter returnType, MediaType selectedContentType,
-                                  Class selectedConverterType, ServerHttpRequest request, ServerHttpResponse response) {
-        if (body instanceof CustomResponse) {
-            CustomResponse<?> customResponse = (CustomResponse<?>) body;
-            HttpStatus status = customResponse.getHttpStatus();
-            response.setStatusCode(status);
-        }
-
-        return body;
-    }
-}
